+++ cypress/support/e2e.js
@@ -18,4 +18,3 @@ import './commands';
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
-require('@replayio/cypress/support');