+++ src/__tests__/useController.test.tsx
@@ -764,4 +764,51 @@ describe('useController', () => {
 
     screen.getByText('pristine');
   });
+
+  it('should disable form input with disabled prop', async () => {
+    const App = () => {
+      const [disabled, setDisabled] = React.useState(false);
+      const { control, watch } = useForm({
+        defaultValues: {
+          test: 'test',
+        },
+      });
+      const {
+        field: { disabled: disabledProps },
+      } = useController({
+        control,
+        name: 'test',
+        disabled,
+      });
+
+      const input = watch('test');
+
+      return (
+        <form>
+          <p>{input}</p>
+          <button
+            onClick={() => {
+              setDisabled(!disabled);
+            }}
+            type={'button'}
+          >
+            toggle
+          </button>
+          <p>{disabledProps ? 'disable' : 'notDisabled'}</p>
+        </form>
+      );
+    };
+
+    render(<App />);
+
+    screen.getByText('test');
+    screen.getByText('notDisabled');
+
+    fireEvent.click(screen.getByRole('button'));
+
+    waitFor(() => {
+      screen.getByText('');
+      screen.getByText('disable');
+    });
+  });
 });