+++ src/types/form.ts
@@ -6,6 +6,8 @@ import { ErrorOption, FieldError, FieldErrors } from './errors';
 import { EventType } from './events';
 import { FieldArray } from './fieldArray';
 import {
+  Field,
+  FieldName,
   FieldRefs,
   FieldValue,
   FieldValues,
@@ -156,6 +158,7 @@ export type KeepStateOptions = Partial<{
   keepValues: boolean;
   keepDefaultValues: boolean;
   keepIsSubmitted: boolean;
+  keepIsSubmitSuccessful: boolean;
   keepTouched: boolean;
   keepIsValid: boolean;
   keepSubmitCount: boolean;
@@ -440,7 +443,7 @@ export type UseFormWatch<TFieldValues extends FieldValues> = {
  * @remarks
  * [API](https://react-hook-form.com/docs/useform/trigger) • [Demo](https://codesandbox.io/s/react-hook-form-v7-ts-triggervalidation-forked-xs7hl) • [Video](https://www.youtube.com/watch?v=-bcyJCDjksE)
  *
- * @param name - provide empty argument will trigger the entire form validation, an array of field names will validate an arrange of fields, and a single field name will only trigger that field's validation.
+ * @param name - provide empty argument will trigger the entire form validation, an array of field names will validate an array of fields, and a single field name will only trigger that field's validation.
  * @param options - should focus on the error field
  *
  * @returns validation result
@@ -777,6 +780,21 @@ export type Control<
   _getFieldArray: <TFieldArrayValues>(
     name: InternalFieldName,
   ) => Partial<TFieldArrayValues>[];
+  _updateDisabledField: (
+    props: {
+      disabled?: boolean;
+      name: FieldName<any>;
+    } & (
+      | {
+          field?: Field;
+          fields?: undefined;
+        }
+      | {
+          field?: undefined;
+          fields?: FieldRefs;
+        }
+    ),
+  ) => void;
   _executeSchema: (
     names: InternalFieldName[],
   ) => Promise<{ errors: FieldErrors }>;